services/ai-service/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
sentence-transformers==2.3.1
numpy==1.26.4
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
openai==1.54.0
google-genai==0.2.0
google-cloud-aiplatform>=1.50.0
